Include recipe resolver logic
