+++ crates/icp-deploy-canister/Cargo.toml
@@ -20,6 +20,7 @@ candid_parser.workspace = true
 dunce.workspace = true
 futures.workspace = true
 glob.workspace = true
+handlebars.workspace = true
 hex.workspace = true
 ic-agent.workspace = true
 ic-management-canister-types.workspace = true
