+++ crates/icp-deploy-canister/src/project.rs
@@ -63,10 +63,17 @@ pub enum ConsolidateManifestError {
     #[snafu(display("failed to load {kind} manifest at: {path}"))]
     Failed { kind: String, path: String },
 
-    #[snafu(display("failed to resolve canister recipe: {recipe_type:?}"))]
+    #[snafu(display("failed to fetch canister recipe: {recipe_type:?}"))]
     Recipe {
         #[snafu(source(from(recipe::ResolveError, Box::new)))]
         source: Box<recipe::ResolveError>,
+        recipe_type: RecipeType,
+    },
+
+    #[snafu(display("failed to render canister recipe: {recipe_type:?}"))]
+    RenderRecipe {
+        #[snafu(source(from(recipe::RenderRecipeError, Box::new)))]
+        source: Box<recipe::RenderRecipeError>,
         recipe_type: RecipeType,
     },
 
@@ -300,17 +307,22 @@ async fn build_manifest_canisters(
                     },
                 ),
 
-                // Recipe
+                // Recipe: fetch the template through the resolver, then render
+                // and parse it into concrete steps.
                 Instructions::Recipe { recipe } => {
                     let ctx = recipe::RecipeContext {
                         canister_name: m.name.clone(),
                     };
-                    recipe_resolver
-                        .resolve_recipe(recipe, &ctx)
-                        .await
-                        .context(RecipeSnafu {
-                            recipe_type: recipe.recipe_type.clone(),
-                        })?
+                    let template =
+                        recipe_resolver
+                            .resolve_recipe(recipe)
+                            .await
+                            .context(RecipeSnafu {
+                                recipe_type: recipe.recipe_type.clone(),
+                            })?;
+                    recipe::render_recipe(&template, recipe, &ctx).context(RenderRecipeSnafu {
+                        recipe_type: recipe.recipe_type.clone(),
+                    })?
                 }
             };
 
@@ -1210,9 +1222,8 @@ pub fn verify_sandbox(project: &Project) -> Result<(), VerifySandboxError> {
 #[cfg(test)]
 mod dependency_tests {
     use super::*;
-    use crate::canister::recipe::{RecipeContext, RemoteResourceResolve, ResolveError};
+    use crate::canister::recipe::{RemoteResourceResolve, ResolveError};
     use crate::manifest::adapter::prebuilt::SourceField;
-    use crate::manifest::canister::{BuildSteps, SyncSteps};
     use crate::manifest::recipe::Recipe;
     use camino_tempfile::Utf8TempDir;
     use tokio::sync::mpsc::Sender;
@@ -1222,11 +1233,7 @@ mod dependency_tests {
 
     #[async_trait::async_trait]
     impl RemoteResourceResolve for PanicResolver {
-        async fn resolve_recipe(
-            &self,
-            _recipe: &Recipe,
-            _context: &RecipeContext,
-        ) -> Result<(BuildSteps, SyncSteps), ResolveError> {
+        async fn resolve_recipe(&self, _recipe: &Recipe) -> Result<String, ResolveError> {
             panic!("recipe resolver should not be called in dependency tests");
         }
