+++ crates/icp/Cargo.toml
@@ -25,7 +25,6 @@ fd-lock = { workspace = true }
 flate2 = { workspace = true }
 futures = { workspace = true }
 glob = { workspace = true }
-handlebars = { workspace = true }
 hex = { workspace = true }
 hmac = { workspace = true }
 hybrid-array = { workspace = true }
