--- crates/icp/src/canister/recipe/handlebars.rs
@@ -1,577 +1,0 @@
-use std::{str::FromStr, string::FromUtf8Error};
-
-use async_trait::async_trait;
-use handlebars::{Context, Helper, HelperDef, HelperResult, Output};
-use indoc::formatdoc;
-use reqwest::{Method, Request, Url};
-use serde::Deserialize;
-use sha2::{Digest, Sha256};
-use snafu::prelude::*;
-use tracing::debug;
-use url::ParseError;
-
-use crate::{
-    fs::read,
-    manifest::{
-        canister::{BuildSteps, SyncSteps},
-        recipe::{Recipe, RecipeType},
-    },
-    package::{
-        PackageCache, cache_registry_recipe, cache_uri_recipe, read_cached_registry_recipe,
-        read_cached_uri_recipe,
-    },
-    prelude::*,
-};
-
-use super::{RemoteResourceResolve, ResolveError};
-use crate::manifest::adapter::prebuilt::SourceField;
-use tokio::sync::mpsc::Sender;
-
-pub struct Handlebars {
-    /// Http client for fetching remote recipe templates
-    pub http_client: reqwest::Client,
-    /// Package cache for caching downloaded recipe templates
-    pub pkg_cache: PackageCache,
-}
-
-pub enum TemplateSource {
-    LocalPath(PathBuf),
-    RemoteUrl(String),
-
-    /// Template originating in a remote registry, e.g `@dfinity/rust@v1.0.2`
-    Registry(String, String, String),
-}
-
-#[derive(Debug, Snafu)]
-pub enum HandlebarsError {
-    #[snafu(display("failed to read local recipe template file"))]
-    ReadFile { source: crate::fs::IoError },
-
-    #[snafu(display("failed to decode UTF-8 string"))]
-    DecodeUtf8 { source: FromUtf8Error },
-
-    #[snafu(display("failed to parse user-provided url"))]
-    UrlParse { source: ParseError },
-
-    #[snafu(display("failed to execute http request"))]
-    HttpRequest { source: reqwest::Error },
-
-    #[snafu(display("request to '{url}' returned '{status}' status-code"))]
-    HttpStatus { url: String, status: u16 },
-
-    #[snafu(display("the recipe template for recipe type '{recipe}' failed to be rendered"))]
-    Render {
-        source: handlebars::RenderError,
-        recipe: String,
-        template: String,
-    },
-
-    #[snafu(display(
-        "sha256 checksum mismatch for recipe template: expected {expected}, actual {actual}"
-    ))]
-    ChecksumMismatch { expected: String, actual: String },
-
-    #[snafu(display("failed to read cached recipe template"))]
-    ReadCache {
-        source: crate::package::RecipeCacheError,
-    },
-
-    #[snafu(display("failed to cache recipe template"))]
-    CacheRecipe {
-        source: crate::package::RecipeCacheError,
-    },
-
-    #[snafu(display("failed to acquire lock on package cache"))]
-    LockCache { source: crate::fs::lock::LockError },
-}
-
-impl Handlebars {
-    async fn resolve_impl(
-        &self,
-        recipe: &Recipe,
-        recipe_context: &super::RecipeContext,
-    ) -> Result<(BuildSteps, SyncSteps), HandlebarsError> {
-        // Determine the template source
-        let tmpl_source = match &recipe.recipe_type {
-            RecipeType::File(path) => TemplateSource::LocalPath(Path::new(&path).into()),
-            RecipeType::Url(url) => TemplateSource::RemoteUrl(url.to_owned()),
-            RecipeType::Registry {
-                name,
-                recipe,
-                version,
-            } => TemplateSource::Registry(name.to_owned(), recipe.to_owned(), version.to_owned()),
-        };
-
-        // Retrieve the template, using cache for remote/registry sources
-        let (tmpl, should_cache) = match &tmpl_source {
-            TemplateSource::LocalPath(path) => {
-                let bytes = read(path).context(ReadFileSnafu)?;
-                (parse_bytes_to_string(bytes)?, false)
-            }
-
-            TemplateSource::RemoteUrl(u) => {
-                // Check cache
-                let maybe_cached = self
-                    .pkg_cache
-                    .with_read(async |r| {
-                        read_cached_uri_recipe(r, u, recipe.sha256.as_deref())
-                            .context(ReadCacheSnafu)
-                    })
-                    .await
-                    .context(LockCacheSnafu)?;
-                if let Some(cached) = maybe_cached? {
-                    debug!("Using cached recipe template for {u}");
-                    (parse_bytes_to_string(cached)?, false)
-                } else {
-                    // Download the template
-                    let tmpl = self.fetch_remote_bytes(u).await?;
-                    (parse_bytes_to_string(tmpl)?, true)
-                }
-            }
-
-            // TMP(or.ricon): Temporarily hardcode a dfinity registry
-            TemplateSource::Registry(registry, recipe_name, version) => {
-                if registry != "dfinity" {
-                    panic!("only the dfinity registry is currently supported");
-                }
-
-                let package = format!("@{registry}/{recipe_name}");
-                let release_tag = format!("{recipe_name}-{version}");
-
-                // Check cache
-                let maybe_cached = self
-                    .pkg_cache
-                    .with_read(async |r| {
-                        read_cached_registry_recipe(r, &package, version).context(ReadCacheSnafu)
-                    })
-                    .await
-                    .context(LockCacheSnafu)?;
-                if let Some(cached) = maybe_cached? {
-                    debug!("Using cached recipe template for {package}@{version}");
-                    (parse_bytes_to_string(cached)?, false)
-                } else {
-                    // Download the template
-                    let url = format!(
-                        "https://github.com/dfinity/icp-cli-recipes/releases/download/{release_tag}/recipe.hbs"
-                    );
-                    let bytes = self.fetch_remote_bytes(&url).await?;
-
-                    (parse_bytes_to_string(bytes)?, true)
-                }
-            }
-        };
-
-        let hash = if let Some(sha256) = &recipe.sha256 {
-            verify_checksum(tmpl.as_bytes(), sha256)?
-        } else {
-            Sha256::digest(tmpl.as_bytes()).into()
-        };
-
-        // Load the template via handlebars
-        let mut reg = handlebars::Handlebars::new();
-
-        // Disable HTML escaping since the output is YAML, not HTML
-        reg.register_escape_fn(handlebars::no_escape);
-
-        // Register helpers
-        reg.register_helper("replace", Box::new(ReplaceHelper));
-
-        // Reject unset template variables
-        reg.set_strict_mode(true);
-
-        debug!(
-            "{}",
-            formatdoc! {r#"
-            Loaded template:
-            ------
-            {tmpl}
-            ------
-        "#}
-        );
-
-        // Build render context: user-provided configuration plus injected _.* variables.
-        // The _ key is reserved and always overrides any user-supplied value.
-        let mut render_context = recipe.configuration.clone();
-        render_context.insert("_".to_string(), recipe_context.to_yaml());
-
-        // Render the template to YAML
-        let out = reg
-            .render_template(&tmpl, &render_context)
-            .context(RenderSnafu {
-                recipe: recipe.recipe_type.clone(),
-                template: tmpl.to_owned(),
-            })?;
-
-        // Read the rendered YAML canister manifest
-        // Recipes can only render build/sync
-        #[derive(Deserialize)]
-        struct BuildSyncHelper {
-            build: BuildSteps,
-            #[serde(default)]
-            sync: SyncSteps,
-        }
-
-        let insts = serde_yaml::from_str::<BuildSyncHelper>(&out);
-        let (build, sync) = match insts {
-            Ok(helper) => (helper.build, helper.sync),
-            Err(e) => panic!(
-                "{}",
-                formatdoc! {r#"
-                Unable to render recipe {} template into valid yaml: {e}
-
-                Rendered content:
-                ------
-                {out}
-                ------
-            "#, recipe.recipe_type}
-            ),
-        };
-
-        // The template is verified good - now cache it if it was remote
-        if should_cache {
-            match tmpl_source {
-                TemplateSource::LocalPath(_) => unreachable!("local files are never cached"),
-                TemplateSource::RemoteUrl(u) => {
-                    self.pkg_cache
-                        .with_write(async |w| {
-                            cache_uri_recipe(w, &u, &hex::encode(hash), tmpl.as_bytes())
-                                .context(CacheRecipeSnafu)?;
-                            Ok(())
-                        })
-                        .await
-                        .context(LockCacheSnafu)??;
-                }
-                TemplateSource::Registry(registry, recipe_name, version) => {
-                    let package = format!("@{registry}/{recipe_name}");
-                    self.pkg_cache
-                        .with_write(async |w| {
-                            cache_registry_recipe(
-                                w,
-                                &package,
-                                &version,
-                                &hex::encode(hash),
-                                tmpl.as_bytes(),
-                            )
-                            .context(CacheRecipeSnafu)
-                        })
-                        .await
-                        .context(LockCacheSnafu)??;
-                }
-            }
-        }
-        Ok((build, sync))
-    }
-
-    /// Fetch raw bytes from a remote URL.
-    async fn fetch_remote_bytes(&self, url: &str) -> Result<Vec<u8>, HandlebarsError> {
-        let u = Url::from_str(url).context(UrlParseSnafu)?;
-        debug!("Requesting template from: {u}");
-
-        let resp = self
-            .http_client
-            .execute(Request::new(Method::GET, u.clone()))
-            .await
-            .context(HttpRequestSnafu)?;
-
-        if !resp.status().is_success() {
-            return HttpStatusSnafu {
-                url: u.to_string(),
-                status: resp.status().as_u16(),
-            }
-            .fail();
-        }
-
-        Ok(resp.bytes().await.context(HttpRequestSnafu)?.to_vec())
-    }
-}
-
-#[async_trait]
-impl RemoteResourceResolve for Handlebars {
-    async fn resolve_recipe(
-        &self,
-        recipe: &Recipe,
-        recipe_context: &super::RecipeContext,
-    ) -> Result<(BuildSteps, SyncSteps), ResolveError> {
-        self.resolve_impl(recipe, recipe_context)
-            .await
-            .map_err(|source| ResolveError::Resolve {
-                source: Box::new(source),
-            })
-    }
-
-    async fn resolve_wasm(
-        &self,
-        source: &SourceField,
-        base_dir: &Path,
-        sha256: Option<&str>,
-        stdio: Option<Sender<String>>,
-    ) -> Result<PathBuf, ResolveError> {
-        crate::canister::wasm::resolve(source, base_dir, sha256, stdio.as_ref(), &self.pkg_cache)
-            .await
-            .map_err(|source| ResolveError::ResolveWasm {
-                source: Box::new(source),
-            })
-    }
-}
-
-/// Handlebars helper for string replacement operations
-/// Usage: {{ replace "from" "to" value }}
-#[derive(Clone, Copy)]
-struct ReplaceHelper;
-
-impl HelperDef for ReplaceHelper {
-    fn call<'reg: 'rc, 'rc>(
-        &self,
-        h: &Helper,
-        _: &'reg handlebars::Handlebars<'reg>,
-        _: &Context,
-        _: &mut handlebars::RenderContext<'reg, 'rc>,
-        out: &mut dyn Output,
-    ) -> HelperResult {
-        let (from, to) = (
-            h.param(0).unwrap().render(), // from
-            h.param(1).unwrap().render(), // to
-        );
-
-        let v = h.param(2).unwrap().render();
-        out.write(&v.replace(&from, &to))?;
-
-        Ok(())
-    }
-}
-
-/// Helper function to verify sha256 checksum of recipe template bytes
-fn verify_checksum(bytes: &[u8], expected: &str) -> Result<[u8; 32], HandlebarsError> {
-    let actual_hash = {
-        let mut h = Sha256::new();
-        h.update(bytes);
-        h.finalize()
-    };
-    let actual = hex::encode(actual_hash);
-    if actual != expected {
-        return ChecksumMismatchSnafu {
-            expected: expected.to_string(),
-            actual,
-        }
-        .fail();
-    }
-    Ok(actual_hash.into())
-}
-
-/// Helper function to parse bytes into a UTF-8 string
-fn parse_bytes_to_string(bytes: Vec<u8>) -> Result<String, HandlebarsError> {
-    String::from_utf8(bytes).context(DecodeUtf8Snafu)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::canister::recipe::RecipeContext;
-    use crate::manifest::recipe::{Recipe, RecipeType};
-    use std::collections::HashMap;
-
-    fn recipe_context(canister_name: &str) -> RecipeContext {
-        RecipeContext {
-            canister_name: canister_name.to_string(),
-        }
-    }
-
-    #[tokio::test]
-    async fn template_values_are_not_html_escaped() {
-        // Create a recipe template that interpolates a value containing
-        // characters that Handlebars would normally HTML-escape (", =, &, <, >).
-        // Use double-stache {{ }} which is what real recipe templates use.
-        let tmp = camino_tempfile::Utf8TempDir::new().unwrap();
-        let tmpl_path = tmp.path().join("recipe.hbs");
-        std::fs::write(
-            &tmpl_path,
-            indoc::indoc! {r#"
-                build:
-                  steps:
-                    - type: script
-                      command: "{{ command }}"
-            "#},
-        )
-        .unwrap();
-
-        let cache_dir = tmp.path().join("pkg");
-        let pkg_cache = PackageCache::new(cache_dir).unwrap();
-        let hbs = Handlebars {
-            http_client: reqwest::Client::new(),
-            pkg_cache,
-        };
-
-        let mut configuration = HashMap::new();
-        configuration.insert(
-            "command".to_string(),
-            serde_yaml::Value::String("SITE=https://example.com&foo=bar npm run build".to_string()),
-        );
-
-        let recipe = Recipe {
-            recipe_type: RecipeType::File(tmpl_path.to_string()),
-            configuration,
-            sha256: None,
-        };
-
-        let (build, _sync) = hbs
-            .resolve_impl(&recipe, &recipe_context("my-canister"))
-            .await
-            .unwrap();
-        let cmd = build.steps[0].clone();
-
-        match cmd {
-            crate::manifest::canister::BuildStep::Script(adapter) => {
-                let commands = adapter.command.as_vec();
-                assert_eq!(
-                    commands[0], "SITE=https://example.com&foo=bar npm run build",
-                    "Template values must not be HTML-escaped (= and & must be preserved)"
-                );
-            }
-            other => panic!("Expected Script build step, got: {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn canister_name_is_injected() {
-        let tmp = camino_tempfile::Utf8TempDir::new().unwrap();
-        let tmpl_path = tmp.path().join("recipe.hbs");
-        std::fs::write(
-            &tmpl_path,
-            indoc::indoc! {r#"
-                build:
-                  steps:
-                    - type: script
-                      command: "build {{_.canister.name}}"
-            "#},
-        )
-        .unwrap();
-
-        let pkg_cache = PackageCache::new(tmp.path().join("pkg")).unwrap();
-        let hbs = Handlebars {
-            http_client: reqwest::Client::new(),
-            pkg_cache,
-        };
-
-        let recipe = Recipe {
-            recipe_type: RecipeType::File(tmpl_path.to_string()),
-            configuration: HashMap::new(),
-            sha256: None,
-        };
-
-        let (build, _sync) = hbs
-            .resolve_impl(&recipe, &recipe_context("my-canister"))
-            .await
-            .unwrap();
-
-        match build.steps[0].clone() {
-            crate::manifest::canister::BuildStep::Script(adapter) => {
-                assert_eq!(adapter.command.as_vec()[0], "build my-canister");
-            }
-            other => panic!("Expected Script build step, got: {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn canister_name_works_with_replace_helper() {
-        let tmp = camino_tempfile::Utf8TempDir::new().unwrap();
-        let tmpl_path = tmp.path().join("recipe.hbs");
-        std::fs::write(
-            &tmpl_path,
-            indoc::indoc! {r#"
-                build:
-                  steps:
-                    - type: script
-                      command: "cp {{ replace "-" "_" _.canister.name }}.wasm out.wasm"
-            "#},
-        )
-        .unwrap();
-
-        let pkg_cache = PackageCache::new(tmp.path().join("pkg")).unwrap();
-        let hbs = Handlebars {
-            http_client: reqwest::Client::new(),
-            pkg_cache,
-        };
-
-        let recipe = Recipe {
-            recipe_type: RecipeType::File(tmpl_path.to_string()),
-            configuration: HashMap::new(),
-            sha256: None,
-        };
-
-        let (build, _sync) = hbs
-            .resolve_impl(&recipe, &recipe_context("my-canister"))
-            .await
-            .unwrap();
-
-        match build.steps[0].clone() {
-            crate::manifest::canister::BuildStep::Script(adapter) => {
-                assert_eq!(adapter.command.as_vec()[0], "cp my_canister.wasm out.wasm");
-            }
-            other => panic!("Expected Script build step, got: {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn reserved_namespace_cannot_be_overridden_by_user_config() {
-        let tmp = camino_tempfile::Utf8TempDir::new().unwrap();
-        let tmpl_path = tmp.path().join("recipe.hbs");
-        std::fs::write(
-            &tmpl_path,
-            indoc::indoc! {r#"
-                build:
-                  steps:
-                    - type: script
-                      command: "build {{_.canister.name}}"
-            "#},
-        )
-        .unwrap();
-
-        let pkg_cache = PackageCache::new(tmp.path().join("pkg")).unwrap();
-        let hbs = Handlebars {
-            http_client: reqwest::Client::new(),
-            pkg_cache,
-        };
-
-        let mut configuration = HashMap::new();
-        configuration.insert(
-            "_".to_string(),
-            serde_yaml::Value::Mapping({
-                let mut m = serde_yaml::Mapping::new();
-                m.insert(
-                    serde_yaml::Value::String("canister".to_string()),
-                    serde_yaml::Value::Mapping({
-                        let mut inner = serde_yaml::Mapping::new();
-                        inner.insert(
-                            serde_yaml::Value::String("name".to_string()),
-                            serde_yaml::Value::String("user-override".to_string()),
-                        );
-                        inner
-                    }),
-                );
-                m
-            }),
-        );
-
-        let recipe = Recipe {
-            recipe_type: RecipeType::File(tmpl_path.to_string()),
-            configuration,
-            sha256: None,
-        };
-
-        let (build, _sync) = hbs
-            .resolve_impl(&recipe, &recipe_context("real-name"))
-            .await
-            .unwrap();
-
-        match build.steps[0].clone() {
-            crate::manifest::canister::BuildStep::Script(adapter) => {
-                assert_eq!(
-                    adapter.command.as_vec()[0],
-                    "build real-name",
-                    "_ namespace must not be overridable by user configuration"
-                );
-            }
-            other => panic!("Expected Script build step, got: {other:?}"),
-        }
-    }
-}
