+++ crates/icp/src/canister/recipe/mod.rs
@@ -1,7 +1,7 @@
 //! Host-side recipe facade.
 //!
-//! The `RemoteResourceResolve` interface + context/error types live in
-//! `icp_deploy_canister::canister::recipe`; the concrete Handlebars resolver
+//! The `RemoteResourceResolve` interface, recipe rendering, and context/error
+//! types live in `icp_deploy_canister::canister::recipe`; the concrete resolver
 //! (which fetches templates and plugin wasms over HTTP and caches them) stays
 //! here.
 
@@ -9,4 +9,4 @@ pub use icp_deploy_canister::canister::recipe::{
     RecipeContext, RemoteResourceResolve, ResolveError,
 };
 
-pub mod handlebars;
+pub mod resolver;
