+++ crates/icp/src/context/init.rs
@@ -3,7 +3,7 @@ use std::{env::current_dir, sync::Arc};
 use snafu::prelude::*;
 
 use crate::canister::build::Builder;
-use crate::canister::recipe::handlebars::Handlebars;
+use crate::canister::recipe::resolver::ResourceResolver;
 use crate::context::Context;
 use crate::directories::{Access as _, Directories};
 use crate::prelude::*;
@@ -89,7 +89,7 @@ pub fn initialize(
     let pkg_cache = dirs.package_cache().context(PackageCacheSnafu)?;
 
     // Recipes
-    let recipe = Arc::new(Handlebars {
+    let recipe = Arc::new(ResourceResolver {
         http_client,
         pkg_cache,
     });
